pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "tdcr-benchmark"
version = "0.1.0"
description = "Python benchmark implementation for tendon-driven continuum robot models."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.22.4,<2.0",
  "scipy>=1.7",
  "matplotlib>=3.5",
  "plotly>=5.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
test = [
  "pytest>=7.0",
]

[project.scripts]
tdcr_benchmark = "tdcr_benchmark.cli:main"

[tool.setuptools.packages.find]
where = ["."]
